This file corresponds to Pekka Sammallahti's file twolrules-saame.txt from 1993. Before cvs inclusion (at least) the following changes were made:
The set DummyNotY5 was added (containing all dummy symbols except Y5),
cf. the following paragraph.

The rule "Postvocalic j Surfacing" has been changed from "<=>" to
"<=", and given a new context ("DummyNotY5", since the old subtraction
set ":0 - Y5:" did notwork). Also, the rule "Postvocalic j in the stem
of words" has been added. As a result, the "mujtal" problem is solved
(the file accepted only j as input for lexical j, but gave correct
output), but certain problems regarding i/j in front of
consonant-initial suffixes still remain. The problem could be due to
the optionality in the rules (the <= operator). This must be solved,
without spoiling the achieved mujtal solution.

The DummyNotY5 set was also inserted to the rule "Word Final Consonant
Neutralization"

The original rule file did not handle the i:a1 alternation of the
boahtit : boad1a1n type. This has been corrected, but I am at the
moment not able to find the correction made.

The following rule, from the original file, has been removed from the
twol-sme.txt file:

"Vowel Shortening in Compounds after Long 1st Syllable 2"
  Vx:Vy <=> [ #: | Cns ] Vz Cns+ Cns: Cns+ _ StemCns:* [ X7: | X8: ] ;
                        where Vz in (a i o u) ;
                        where Vx in (u i)
                                Vy in (o e)
                        matched ;

This is (probably) an error, and it has tentatively been included
again. The rule "Optional Shortening of a1 in Compounds after Long 1st
Syllable 2" was found duplicated in the cvs 1.1.1.1 version and the
duplicate removed (evidently, the wrong rule was moved).

Some of the two-part rules were found not next to each other in the
original rule file, they are now adjacent to each other. As a result,
the rules "Optional Vowel Shortening after Short 1st Syllable" and
"Optional Vowel Shortening in Caritives" now are found earlier in the
rule file than before. The rules are not changed, though.

The word "matched" was added to the rules "Stem Vowel in Imperative
Plural 1st Person and Passive Stem" and "Word Final Deletion of n2 m2
g2 h2". Also, the "DelCns" of the latter rule was put in parentheses.