This is a major revision of the tok.txt file, aimed at giving a better treatment of numerals. Until now, NUM has been a string of numbers and number operators or seperators, in any order. Thus, no sentence-final numerals have been possible, as the final period has been seen as a part of the numeral.
The present version changes that, by redefining NUM as to having a
number as its final element, and thereafter by treating numerals as
abbreviations (if the following word starts in a capital letter, copy
the punctuation mark to the next line, otherwise leave it as part of
the numeral (ant treat it as an ordinal).

A new tokentype, TOKNUMOrd, was introduced, referring to NUMOrd,
itself a sequence of NUM and %., and the treatment of TOKNUMOrd was
done dependent upon the following capital.

Finally, a good deal of the abbreviations were tempoarily commented
out, since the compilation time proved to be long. TODO: Evaluate
abbreviations.