The lexicon ASEHIS was removed, and all ASEHIS in the lexicon redirected to ASEHAS. Then, the content of the two former lexica were merged into one ASEHAS lexicon. Lexica formerly pointing to ASEHAS got the content of the old ASEHAS spelled out. The reason for this was that there were errors in the old share of burden. BOARIS should go as odd, not as even, and there should be no need to cross-refer to the old ASEHAS lexicon.
A new RAHKIS lexicon was introduced, giving a facultative -is
attribute in addition to the -es attribute. A way of restructuring the
adjectival lexica could be to look more systematically into a division
between attribute form and the rest, so that all lexica point to
follow-up lexica, save the attributive form.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@1081 c7155fb1-f0a7-4240-a2fc-2600b6f42f90