Added LEX-{N, V, A} sets to denote lexicalised words. These were then used in rules to prefer non-derived words. Note that the rules in question should be flipped around in case it is interesting to see the structure of the words.
Today's news was the trivial (well, when you know it) observation that
you may ahve sets as target without parentheses, and tags as targets
#with# parentheses. So, the relevant rules were rewritten and
included.

Rules for selecting Prop were rewritten.

The Px rules were rewritten, as follows: If the Px candidate meets the
subj - V - self pattern, choose Px, if not, delete Px. Thus, most Px-s
will be deleted. It is still an open question as to whether it is
enough.

The rules for Inf selection were simplified and partly also
generalised (referring to less specific contexts). Generalisation is a
good idea, it must be pursued more.

The Attr selection / removing rule set was made more active, the
philosophy being that the Attr + Noun is a strong relation. If there
is no candidate to the right, then you are no attribute, and if there
is an attribute looking for y9ou, then you are a noun.

All the rules for DEM and VFIN selection had been out of order due to
the SET TARGET without () misunderstanding. They were now included.