This is a major change. The whole file is reshaped with this check-in. Rather than grouping the rules around the disambiguation patterns, the rule follow the grammar, we can thus go from category to category and find rules that select the relevant categories.
The reason for this is that the open issues are more clearly visible,
and it is easier to find one's way in the file.

Also, several new rules have been added. I have started defining
complementary sets (NOT-PRE-NP-HEAD etc.), the intention is to use the
BARRIER formalism more efficiently.

Several rules were written for verb disambiguation, and some of the
accusative rules were sharpened as well.