Got rid of \201 and \216 in the source file, now sets with á are possible. Revised the sets: Added Neg to finite verbs, made more syntactic, complementary sets, for the use with barriers.
Inserted several CONSTRAINT and SETS symbols, this to avoid a bug in
the program: It starts malfunctioning when there is too many sets and
rules between the constraint and set symbols.

Worked especially on predicatives and on Pl3 disambiguation. Made
several rules that used the BARRIER - LINK formalism, removed earlier,
non-flexible rules. Moved the rules for disambiguating adjective
predicaties to the chapter for nouns, for syntactic (if not POS)
concistency.

Stopped using the "<wordform>" SELECT IF ... formalism. The reason for
this is that it is not case insensitive in the same way as the (0
("lemma")) notation is.