SELECT Gen IF (-1 NP-BORDER)(*1 N BARRIER NON-A LINK 1 VFIN); # hmm SELECT Gen IF (-1 NP-BORDER)(*1 N BARRIER NON-A LINK *1 VFIN); # hmm
Changed former version to latter, because of examples like
"Sápmelac1c1aid historjá álbmogin lea duháhiid jagiid boaris". The
"álbmogin" part had intervened, and the rule is not more
general. There is no barrier, though. Added the following rule to the
same clause:

REMOVE (A Gen) IF (0 N)(*1 N BARRIER NON-A);

The idea is that adjectives do not show case agreement, thus
"sápmelac1c1aid" must get a nominal interpretation.

Added "BARRIER CS" twice to the following rule, in order to keep verbs
within their own clauses. I did not include COMMA in the barrier, but
it is clear that the barrier must be refined.

REMOVE Acc IF (NOT *-1 V-NON-AUX BARRIER CS)(NOT *1 V-NON-AUX BARRIER CS);

Merged the two non-coordinaded accusative object rules into one:
SELECT Acc IF (*-1 V-NON-AUX BARRIER NPNHA)(0 NP-HEAD)(NOT 0 PrfPrc)(1 CLB);

Looked into the Sg1 ~ PrfPrc issue. Until now, the two rule chapters
have not used the star notation; the rules have been
non-flexible. Now, the rules were made according to the following
philosophy: We have Sg1 if there is a Sg1 Pron, or if there is no
finite verb at all. The PrfPrc chapter was moved to the beginning of
the verb chapter, and terminated by a CONSTRAINTS symbol, so that the
parser first picks it PrfPrc when there is a finite verb to link onto,
and then subsequently the Sg1 chapter gives us Sg1. The PrfPrc should
have had an addition in order to handle coordination of two Sg1 ("Mun
human ja human"), this is still not in place.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@1258 c7155fb1-f0a7-4240-a2fc-2600b6f42f90