The sets: Included ToC. Removed sets and traces of the Adp tag in the rules (the tag is now replaced with Pr and Po, the rules are not checked yet. PUNCT is added and used. VFIN has excluded ConNeg as one of its possible members, instead there is an intermediate set V-MOOD used for just the set of verbs conjugated in mood. A new set NPNHAII was born, including the set of illatives on the long subtracted list, thereby opening up for illative 'indirecte objects'.
The rules:

The adverbs section was organized, with single word rules
arphabetically.  The Acc and Gen rules got better borders and
barriers. The Prs Du1 / Prt Pl3 rule set was reshaped, moving a rule
from the latter to the former, and removing CC as a criterion (that
was a mistake earlier). NB! due to an interruption I now see that th
rule referred to with "Anne ja Biret c1ilge, ahte dat lea duohta." is
wrong, it selects Du1 but refers to Du3 subject. I was so sure I had
corrected it :-(. Do this tomorrow! Infinitive rules were improved.,
added an Inf IF -2C Inf -1 CC rule, and making better barriers on old
rules. Rules were added to distinguish between Pr and Po, and a rule
set was added for the GASKAL cases (p-positions that take coordinated
NPs as complements.