Done numerous changes (puh). Added sets that have been taken into use (Refl, PrsPrc, given GASKAL and VAHKU new members, made new complement sets.
Moved the adposition and subjunctions chapters to the beginning of the
file. Since these disambiguation rules are local in nature, I wanted
them more to the front. I also refined the Po rules, a.o. demanding a
preceeding genitive noun (if a noun is present, that is).

Changed more contexts than rules, both for adjectives and nouns
(Especially the genitive and accusative sections). Made more rules for
illative/sg3. I added a new CONSTRAINTS border in the beginning of the
Genitive section, in order to prevent errouneous parsing of genitive
strings.

Weakened the ConNeg rule, now allowing for a PrfPrc barrier to the Neg
(earlier only VFIN barrier). Did a lot of work with the infinitive rules.

Repaired some weaknesses with the "mii" rules. Earlier, they had only
specified the 0 context to be "mii", this errouneously included the
form "maid" in the set as well. Now, these rules require "mii" Nom,
and the maid disambiguation is taken care of in the Adv section.

Made rules for "dasa".


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@1301 c7155fb1-f0a7-4240-a2fc-2600b6f42f90