Added two new sets, for verbs that take infinitives as complements. The sets were taken directly from Nickel. Added a new set NPNHAIIP, that subtracts particles in addition to the earlier stuff (used in the main accusative rule).
Moved the whole pronoun chapter to the beginning. Should consider
moving the verbs vefore the nouns as well.

Relaxed a rule for genitive after numerals, now checking whether the
numeral is a complement to a preceeding preposition. Added the P in
NPNHAIIP, allowing particles to come between a verb and its
object. For the other Acc rules I relaxed their right context.

The infinitive sctions got updated ith a rule using the INF-VERB set
(the ACC-INF-VERB set is still unused).


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@1305 c7155fb1-f0a7-4240-a2fc-2600b6f42f90