Added the atomic sets Du, Cmpnd, as1. Made segs V-DU and V-PL.
Added rule for as1 adjective inf CC to the left and Cmpnd to the
right. This to avoid diminutive nouns in a context where there may be
adjectives, adverbs and perhaps genitives (hmm, could have added
proviso permitting genitive).

Added rules for emphatic reflexives ies1 and iez1a. (hence the set
Du). Added a rule saying that "dasa" is always a pronoun, and never a
VGen of the verb "dassat". A set of 4 earlier dasa rules were thereby
retired, but kept in the file in case VGen forms of "dassat" should
pop up. Made a rule for "nubbi", letting it be Card after "okta" but
Indef after "nubbi", modeled after Finnish "ensimmäinen - toinen" and
"toinen - toinen".

Added a C to the *-1 requirement on the V-NON-AUX in the major
postverbal accusative rule. The philosophy is that when a verb is
prominent enogh to serve as a transitive verb in a sentence, we should
have been able to identify it as such. Words with multiple readings
V/N are often nouns derived from verbs, or they have alternative V/N
readings that make it unlikely that they take objects.

The main news for this check-in is the new multipart compound rule for
locative/comitative disambiguation. The new rule format handles "A and
B- and C" -contexts (i.e., contexts, where the rightmost CC sister
itself is split in two by a CC, and marked with dash (thus tagged with
Cmpnd). This format should be extended to all CC rules.