This is a major revision, containing 3 larger revisions and some smaller.
First, I have inserted word pairs for testing. Positive pairs (pairs
intended to work) are marked with "!â¬", and negative pairs (pairs made
in order to make the parser fail) are marked with "!$". Each part has
one upper and one lower part, these must have the same number of
symbols. Not all rules have test pairs, but the intention is that we
write pairs for each rule. There are positive pairs for most rules,
but negative pairs for only a couple of rules in the beginning.

Note that not all the positive pairs are accepted by the parser at the
moment (11 are not, whereas 76 are, and there is one feasible pair
error). The error is probably in the pairs and not in the parser, but
they should in any case be revised.

Then, I have (with Sjur) written a lengthy introduction in the
beginning. This intro is to long, it will be moved to the
documentation section when it is more stable. Partly, it now functions
as a spec for the programmer to make the necessary scripts.

Third, this is the cvs checkin of our work at the Helsinki meeting in
august. The following rules are simply commented out:

"Optional Vowel Shortening after Short 1st Syllable"
"Vowel Shortening in Vowel-Final Compounds after Long 1st Syllable 1"
"Vowel Shortening in Vowel-Final Compounds after Long 1st Syllable 1"
"Compulsory Vowel Shortening after ij"

As far as we could observe in Helsinki, the parser worked just as fine
without these rules. Now, we should include the test sentences from
these rules into our test suite and evaluate whether this really is
the case.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@4466 c7155fb1-f0a7-4240-a2fc-2600b6f42f90