This is the revision where I introduce our 6 semantic tags +Fem, +Mal, +Sur, +Obj, +Org, +Plc. It was done in the following way: Each BERN, LONDON, etc. lexicon has been split in 7 in the lexicon (6 sem-tags -plc etc. plus no-tag), these now got separate lexica, that added the tags +N+Prop+Fem (for -fem, and correspondingly for the rest). The suffixlex lexica LONDON etc. did not add semantic tags. All these seven lexicon varieties point to an eitht lexicon BERNDECL, LONDONDECL, etc., which corresponds to the old lexicon BERN, LONDON etc. The subsequent sublexica are as before, except the tags +N+Prop were removed, since they had to be inserted #before# the semantic tags, to get a string +N+Prop+Plc+Nom+Loc, etc.
There is a possible problem lurking in this update: In some cases,
there were hyphens and stuff inserted before the +N+Prop tags (%-), I
could not quite see why it was done that way. This was ignored in this
update, and the hyuphens etc. left where they were. The fact that they
are on a different level may give the desired result, despite the
rearrangement, it remains to be seen, but the issue deserves
attention.

Then, all DEATNU entries were moved to MARJA-plc, as the file stated
that all DEATNUÆ entries were placenames. Also, a number of other Sámi
names got tags, on the basis of the final part of the name compound.

The file should now be ready for a final semantic tagging. Note that
new lexicon-suffix combinations must be declared as they are tagged,
it cannot be done on beforehand, since the compiler would
complain. Many of them are made ready for use (just remove the initial
exclamation mark), but others must be manually written along the way.