Several minor updates.
Added set COUNTABLE, for oassi, kapihttal, etc., and made rule for
selecting Num after COUNTABLE.

Made rule for alit oahpu etc.

Made rule for Dem Acc in front of unambiguous Acc noun.

Exchanged NOT-ADV-PCLE-NEG barrier for a Com/Loc rule with a positive
(@SUBJ) OR (@OBJ) OR (@SPRED) OR (@OPRED), hoping it is strong enough,
still permissive enough to accept "earret earÃ¡".

Added rule SELECT (V Pl3) IF (-1 CS LINK -1C (V Pl3)) for subj-pl3
V-pl3 go V-pl3.

Finally added some tidying rules at the end: If a naked acronym or
Numeral have not been disambiguated to Gen or Acc during the race,
they are settled as Nom in the final round. The logic is that if they
don't have the context to be Gen or Acc, we take them to be stray
elements, and offer Nom as the best we can do.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@5051 c7155fb1-f0a7-4240-a2fc-2600b6f42f90