Removed case requirement from < MAP (@NumN<) TARGET Num IF (NOT 0 Nom OR Acc OR Gen OR ADVLCASE); this rule had been dysfunct since we introduced case on numerals. Changed the rule into a rule assigning @NumN< to numbers following N, preceding EOS.
For the same context, I revised a rule marking N + Num sentence finally as Acc +
Nom, rather than Gen + Acc (for strings like "Oainnán diedáhusa 4". Cf: before/after:
<       (*1C Acc BARRIER NPNH OR PrfPrc OR CLB OR CC OR PUNCT LINK 1 EOS);
>       (*1C (Acc N) BARRIER NPNH OR PrfPrc OR CLB OR CC OR PUNCT LINK 1 EOS);
Since the new version requires the Acc to the right to be noun, "4"
cannot be seen as the Acc candidate. Both Num, Pron and A are unlikely
targets of modifying genitives, so this should work in many cases.