Lindas christmas checkin:
Worked with the mapping of numerals.
Added the rule to get rid of the @X mapping for the numeral (interpreted the example sentence as synonymous to "FÃ¡ksa lea...")
>MAP (@SPRED) TARGET Num IF (-2 N)(-1 (":"))(0 CARDINALS)(1 CLB);
    ## FÃ¡ksa: 22242786.

Changed the following rule by concentrating on the missing finite verb and making sure that only the first Nom entity gets this mapping:
>MAP (@HNOUN) TARGET (Num Nom) IF (-1 BOS)(1 EOS);
> SELECT (@HNOUN) IF (NOT -2 (N Nom) OR (ACR Nom))(-1 ("nr") OR ("nummar") OR ("nummir"))(NOT *-1 VFIN OR ("leat" PrfPrc))(0 (Num Nom))(NOT *1 VFIN);
>     ## DieÄÃ¡husa nummar okta.9625a9665,9666> SELECT (@HNOUN) IF (NOT *-1 VFIN OR ("leat" PrfPrc) OR (N Nom) OR (ACR Nom))(0 (N Nom))(NOT *1 VFIN);
>     ## Oassi II.

Added (NOT 0 ("nummar") OR ("nummir))(1 CARDINALS) to the rule:
>MAP (@HNOUN) TARGET (N Nom) IF (NOT *-1 VFIN OR ("leat" PrfPrc) OR CP)
    (NOT 0 ("nummar") OR ("nummir))(1 CARDINALS)
	(NOT *1 (N Nom) BARRIER CS OR CC)
	(NOT *1 VFIN BARRIER CS)(NOT 1 go OR COMMA);

Changed the rule by replacing (1 EOS by ..OR (N Nom) OR (ACR Nom))(NOT 0 ("nummar") OR ("nummir") LINK 1 CARDINALS)(NOT *1 VFIN):
>MAP (@HNOUN) TARGET (N Nom) IF (NOT *-1 VFIN OR ("leat" PrfPrc) OR (N Nom) OR (ACR Nom))(NOT 0 ("nummar") OR ("nummir") LINK 1 CARDINALS)(NOT *1 VFIN);

Mofidied the rule in terms of allowing "nr":
>SELECT Num IF (-1 COUNTABLE OR ("nr"))(0 ABBR OR ACR) ;
	## DieÄÃ¡husa oasis I ovdanbuktojuvvo Bb:a sÃ¡mepolitihkalaÅ¡ vuoÄÄolÃ¡vdi.
This one needs to be modified further.

Added the rule to avoid acronyms such as II after a headnoun:
>SELECT Num IF (*-1 (@HNOUN))(0 (Num Nom) OR (Num Attr));

Modified the rule by adding ACR and (NOT -1 ("nr") OR ("nummar") OR ("nummir")):
>MAP (@NumN<) TARGET Num IF (-1 N OR ACR)(0 CARDINALS)(1 EOS)
    (NOT -1 ("nr") OR ("nummar") OR ("nummir"));

Added the rule:
>MAP (@NumN<) TARGET Num IF (-1 N OR A OR ACR)(NOT -1 ("nr") OR ("nummar") OR ("nummir"))(0 CARDINALS)(NOT 1 N OR A OR ACR);

Added:
>MAP (@NNum>) TARGET N IF (0 ("nummar") OR ("nummir"))(1 WORD OR Num);

Added the rule in order to get nominative for nummar and nummir preceding a cardinal:
>REMOVE (N Gen) IF (0 ("nummar") OR ("nummir"))(1 CARDINALS OR WORD);

Modified the rule by taking away (NOT 0 (Num Nom) LINK NOT -1 (BOS) LINK NOT 1 (EOS)):
>MAP (@SUBJ @SPRED) TARGET Nom IF (NOT 0 ("diibmu") OR ("dbm") LINK *1 Num BARRIER NP-BORDER)
	(NOT 0 Num LINK *-1 ("diibmu") OR ("dbm") BARRIER NP-BORDER) (NOT 0 (Num Nom);

Added (NOT -2 Gen) to get this rule to avoid wrong mappings for examples such as "DieÄÃ¡husa oassi I.":

>MAP (@NumN<) TARGET (Num Nom) IF (-1 ("nr" Nom))(0 CARDINALS)(-2 N OR A OR ABBR OR ACR)(NOT -2 Gen);

For disambiguation of locative and comitative, I tried:
>REMOVE Com IF (-1 (N Ill) OR (ACR Ill));

>REMOVE Com IF (*-3 (N Ill) OR (ACR Ill))(-1 CC OR COMMA)(-2 Com);
    ## DoarjjaruÄaiguin SÃ¡mediggi sÃ¡httÃ¡ vuoruhemiidis mielde addit doarjaga
    ## kvalitehtaloktendoaimmaide gielddain, fylkkagielddain ja oahppoÃ¡sahusain.

Those rules have to be modified further.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@5496 c7155fb1-f0a7-4240-a2fc-2600b6f42f90