Invalidated  #SELECT (PxSg3) IF (-1 (V TV Sg3) LINK NOT 0 Pass LINK *-1 ("son" Sg3 Nom) OR (N Sg Nom) BARRIER NOT-ADV-PCLE) 	#(NOT 0 N-NOPX)(NOT 0 Attr LINK *1 N BARRIER NPNH);
because it errounously choses PxSg3 in some cases.

Added: (NOT -1 @GN> LINK 1 ABBR)
to the following rule:
REMOVE Nom IF (*-1 BOS OR CS OR REALCLB BARRIER NOT-NP)(NOT 1 Prop)
	(*1C Nom BARRIER NPNH OR CC OR CLB OR PUNCT LINK *1 VFIN BARRIER NOT-NP)
	(NOT 1 (Interr Nom))(NOT -1 @GN> LINK 1 ABBR);

So that Â§ "DearvvaÅ¡vuoÄabargiidlága Â§ 10" keeps the nominative tag.

Added the following rule to fix the tag for Â§ in front of numerals:
REMOVE Gen IF (-1 (@GN>))(0 ("Â§"))(1 Num);