Added a new set NOT-A, with obvious content.
To the 1244 rule
MAP (@GN>) TARGET Gen IF (*-2 ("s") OR ("siidu") OR ("Â§") OR ("paragrÃ¡fa") OR
    ("kapihttal") OR ("kapihtal") OR ("oassi") OR ("jahki") OR GEASSET LINK 1
    CARDINALS)(*1 N BARRIER NOT-A);
I added the demand that the first word to the right should be a noun, we got hits giving the @GN> tag when no N to the right was present.

Added a new rule dealing with the INSTITUTION set. In the first
version, "Oslo universiteahtta" and the like came out as nominative,
since nominative and genitive are identical. I added a "C" to the (0
Nom) of the original 5903 rule (giving @PROP> only to unambiguous
nominatives, and added a further 5906 rule giving @GN> to genitives in front of institutions. As a last minute insight I see that the "Prop" tag is missing from the 5906 rule, it should be added.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@6147 c7155fb1-f0a7-4240-a2fc-2600b6f42f90