Added "..." as DELIMITER, we thus now have a larger delimiter set.
Sharpened the special proper name rules for Dan and Dien (4924ff),
demanding them to be in the nominative in order to be hit by the rules (since they
only here are homonymous with the pronouns. Also added a rule for sentence-initial
"Lea", trying to difference between the woman and the copula.

Added a left condition to the 5172 rule on A Attr selection in front of nouns,
excluding predictatives with sentence-initial copulas from the set, and also demanding
a noun to the right. New rule:
> SELECT (A Attr) IF (NOT *-1 LEAT-FIN-NOT-IMP BARRIER NOT-ADV LINK *-1 BOS BARRIER NOT-ADV)
> 	(NOT 0 Adv OR Inf)(1 A-CASE LINK 1 CLB  LINK 1 NP-BORDER);

For a set of time adverbial Acc rules (7704-8890), I added a demand on @ADVL as well,
thereby dismissing the Acc @OBJ readings (from "SELECT Acc IF..." to
"SELECT (Acc @ADVL) IF...".

Small changes, but improved results.