Wrote another constraint to the following rule that is supposed to exclude expressions in the list TIME that follow or precede a transitive verb from the @ADVL tag:
SELECT (Acc @ADVL) IF (-1C NP-BORDER OR A)(NOT -1 Interr OR Indef OR Pr)(NOT 0 Loc)
	(0 TIME)(NOT *1 TIME BARRIER NPNH)(NOT 1 ("loahppa") OR ("álgu"))
	(NOT -1 TV LINK NOT *-2C (@OBJ) BARRIER CS OR CP LINK NOT *1C (@OBJ) BARRIER TV OR CP OR CS);

(7807)

Please check this rule, Trond, and have a look at the barriers.

parallel to that rule comes this one later (it would have been nice if it stood before it):
SELECT (@OBJ) IF (-1 TV)(0 TIME OR ROUTE)(NOT *-2C (@OBJ) BARRIER CS OR CP)(NOT *1C (@OBJ) BARRIER TV OR CP OR CS);

(10783)

Those rules should be tested.