Made yesterdayÂ´s numeral rules stricter.
removed:
REMOVE (N Ord) IF (-1 NP-MEMBER)(NOT 1 NP-MEMBER);

 Added:

SELECT (A Ord) IF (-1 Prop LINK NOT 0 Plc);
	## Heinrich IV. lea boaris.
	## Heinrich njealjÃ¡t lea boaris.

MAP (@APP) TARGET (A Ord) IF (-1 Prop LINK NOT 0 Plc);

Added the last constraint "(NOT 0 (A Ord) LINK -1 Prop LINK NOT 0 Plc)" to:

REMOVE A IF (NOT -1 ("mii" Sg Gen))(0 N)(NOT 0 Comp OR Ess OR ("veara") OR
	("mearkkaÅ¡an#veara") OR ("namuhan#veara"))(1C VFIN)
	(NOT 1 ("leat") LINK *1 ("ahte") BARRIER NOT-ADV-PCLE)
	(NOT 1 go LINK *1 COPULAS BARRIER V-NOT-AUX  LINK *1 (@SUBJ) BARRIER NOT-NP)
	(NOT 0 Qst LINK *1 COPULAS BARRIER V-NOT-AUX  LINK *1 (@SUBJ) BARRIER NOT-NP)(NOT 0 (A Ord) LINK -1 Prop LINK NOT 0 Plc);


Added the last constraint to the following rules:
REMOVE A-CASE IF (*-1 BOS OR CS OR (":" CLB) BARRIER NOT-DEM)(0 N)(NOT 0 ("buorre"))(1 N)
(NOT 0 Ord LINK *1 N BARRIER NOT-A);


MAP (@AN>) TARGET (A Ord) IF (*1 N BARRIER NOT-A);

made a list:

LIST SPATIAL = "alu" "asa" "gasu" "govddu" "guhku" "viidu";
    # Those are words that only appear in one form, which Nickel calls A/G.
    # They are only in a predicative position.

Added the set
SELECT Acc IF (0 Num LINK *1 MEASURE BARRIER NOT-A LINK 1 SPATIAL);
SELECT Acc IF (0 Num LINK 1 SPATIAL);
SELECT Acc IF (0 Num LINK 1 Num LINK 1 SPATIAL);
SELECT Acc IF (0 SPATIAL);
early in the Accusative case chapter. They do not work as intended, though.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@6678 c7155fb1-f0a7-4240-a2fc-2600b6f42f90