Added "dearvvaÅ¡vuoÄabargii#láhka" to LAHKA.
Added the rule
REMOVE V-1-2 IF (*-1 INSTITUTION LINK 1 Po LINK 0 ("bealis") OR ("beale") BARRIER NOT-ADV)(0 OPRED-VERB LINK 0 V-3)(*1 Ess BARRIER NPNHA);
Initially, we had problems with the use of sets V-1-2 and V-3, but it actually works, cf. the 3rd word of the example sentence:
"<atnet>" S:6020, 10286
        "atnit" V TV Ind Prs Pl3 S:1702 @+FMAINV
where "10286" refers to the rule in question.