REMOVE Loc IF (*-1 BOS LINK NOT *1 ILL-VERB OR ("leat"))(0 (Sg Com) LINK *-1 Ill BARRIER CP OR CS); REMOVE Loc IF (*-1 Ill)(*-1 BOS LINK *1 ("gullat") OR ("leat"))(0 (Sg Com) LINK *-1 Ill BARRIER CP OR CS);
Those rules are supposed to hit those cases where both illative and locative exclude each other. This is probably NOT the case in illative verb constructions where the illative does not denote a "location (goal)", with habeo-constructions, where the locative gets another function, and probably not with partitive constructions. Unfortunately, those are not mentioned in the rules.

This rule should be tested.

Added a couple of profession terms from risten.no to the LIST PROFESSION


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@6896 c7155fb1-f0a7-4240-a2fc-2600b6f42f90