Added áÅ¡Å¡i to NOM-GEN-NOUN set. Moved the two rules under the heading "Some particular subjunctions from the beginning to the end of the subjunciton chapter, they were in the way. Now, the first of these  is special and housld stay, but I suspect the second one does not hit any more.
Added a new rule > REMOVE Prop IF (0 CNOUN)(0 (Prop Pl)); in order to chose common nouns
over propernames when in plural.

Made a new rule for catching topicalised acc object, the rule is relevant only for
Nom=Acc=Gen nouns, for all other nouns the existing rules will do. the new rule is:

> SELECT Acc IF (*-1 BOS BARRIER NPNHA OR Nom)(*1 V-SG3 BARRIER NOT-NOM)
> 	(*2 TV BARRIER Acc LINK *1 EOS OR CS OR CLB BARRIER Acc) ;
> 	## Dan áÅ¡Å¡i ferte čilget nállevealaheami ektui sierra.
> 	# This rule is relevant for nouns where Nom=Acc=Gen.

In connection to that, I also changed two "Select Nom" rules and one "Remove Acc" rule,
in order not to bleed the correct accusative analysis for the new SE Acc rule. The
addition involved chccking for Acc to the right (if not - go for a topicalised object),
as well as making better barriers. The rules were (< = old and > = new, as always):

< SELECT Nom IF (*-1 CS OR BOS OR CC OR CP OR PUNCT OR (":") BARRIER NPNHA)(0 Sg)
> SELECT Nom IF (*-1 CS OR BOS OR CC OR CP OR PUNCT OR (":") BARRIER NPNHA OR (Dem Acc))(0 Sg)
< SELECT (N Sg Nom) IF (1 V-SG3)(NOT 1 N)(NOT *2C Nom BARRIER CS);
> SELECT (N Sg Nom) IF (NOT 0 Acc)(1 V-SG3)(NOT 1 N)(NOT *2C Nom BARRIER CS);
< REMOVE Acc IF (0 (N Sg Nom))(1 V-SG3)(NOT *1C Nom BARRIER CS);
> REMOVE Acc IF (0 (N Sg Nom))(1 V-SG3)(NOT *1C Nom BARRIER CS)(NOT *1 TV LINK *1 EOS OR CS OR CLB BARRIER Acc);