Added "áÅ¡Å¡i" to KLASS (the ÃÅ¡Å¡i 14/05 phenomenon).
Made mapping rule to add @APP to nominative numerals following KLASS. Oops, I added the version that did not require the following word to be a Prop. This rule must be monitored.

Added a rule
> MAP (@-FMAINV) TARGET PrfPrc IF (-1 COMMA)(NOT *-2 VFIN)(NOT *1 VFIN);
for elliptic participles. This could be an adjectively used participle, the rule thus have to be altered in order to avoid those cases.

Added a further sigh to the SUBSTITUTE section.