Invalidated the first rule and made two others:
#MAP (@PrcN>) TARGET PrfPrc IF (*-1 V-NOT-AUX BARRIER CS OR CP OR AUX
#	LINK NOT 0 A OR N)(0 Pass)(1 N)(NOT 1 TIME);
	## Vuoigatvuohta dearvvaÅ¡vuoÄaveahkkÃ¡i gusto riikka almmolaÅ¡ ruhtaduvvon
	## spesialistadearvvaÅ¡vuoÄabÃ¡lvalussii.

MAP (@PrcN>) TARGET PrfPrc IF (*-1 V-NOT-AUX LINK NOT 0 PrfPrc BARRIER CS OR CP OR AUX
	LINK NOT 0 A OR N)(0 Pass)(1 N)(NOT 1 TIME);

MAP (@PrcN>) TARGET PrfPrc IF (*-1 CS OR CP BARRIER AUX LINK NOT 0 A OR N)(0 Pass)(*1 N LINK NOT 0 TIME BARRIER NPNH);

The first one was modified with respect to the element to the left necessary for the mapping. The element is not supposed to merely be a coordinated PrfPrc.

The next rule is responsible for cases in which there actually is a CS or CP to the left but no auxiliary after that, including also the possibility of a coordinated @PrcN>.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@7548 c7155fb1-f0a7-4240-a2fc-2600b6f42f90