Changed SELECT Pers to REMOVE DEM in the rule < SELECT Pers IF (0 ("dat"))(NOT 0 ("dasgo") OR ("danin") OR ("dainna")) in order not to remove the Prop in "Dan boahtÃ¡."
Added "(NOT 0 Dem)" to the Px rule
> SELECT (PxSg1) IF (-2 MUN)(-1 V-SG1)(NOT 0 Dem)(NOT -1 ("leat"))
in order not to analyse "duon" as PxSg1 of "duo". (at that stage, "duon" is still seen as either Ess of Dem "duo" or as PxSg1 of the noun "duo".

Commented out the rule
> #REMOVE Dem IF (0 Ess)(NOT *1 (N Ess) BARRIER NPNH); # duon
It actually did good only to the demonstrative "danin" (cf. the example sentence
<       ## DoaibmaplÃ¡nas deattuhuvvojit danin doaimmat bargiid hÃ¡rrÃ¡i.
and for that sentence we had a rule already.
As it stood, the rule only removed Dem readings of duo (duon), and probably it did other inappropriate things as well.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@7751 c7155fb1-f0a7-4240-a2fc-2600b6f42f90