Changed < REMOVE Actio IF (*-1 NP-BORDER BARRIER NOT-PRE-A-N)(*1 VFIN) ; to > REMOVE Actio IF (-1 PRE-A-N)(0 PrfPrc)(1 N); The previous version was simply incomprehensible, it somehow had a failed logic. Now, we reason like this:
In order for a participle (and yes, they are participles, so (0
PrfPrc)), it needs to have a noun to the immediate right, and never
mind the VFIN, it could be anywhere, if at all. To the left, we err on
the safe side, and ask for a PRE-A-N, that is some Dem. The existence
of NP-BORDER is irrelevant. Later, we could consider removing the -1 requirement.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@7752 c7155fb1-f0a7-4240-a2fc-2600b6f42f90