Made a rule for verbal compounds:
REMOVE ("olggo#saddit") IF (0 ("olggos#addit"));

Changed duo to du^o in several rules


Wrote a modified versjon of the previous SELECT Acc rule just for "duon", so that we wonÂ´t get the Essiv inside an NP, where duon clearly is a demonstrative:

SELECT Acc IF (*-1 V-TRANS-ACT BARRIER VERB OR CC OR CLB OR CS OR (@OBJ)
	OR PUNCT OR ("dego") LINK NOT 0 AUX OR (@ADVL) OR PrsPrc)(0 ("duot"))(NOT 0 TIME)
	(NOT 0 (Acc PxSg3) LINK 0 Loc)
	(NOT 0 Nom LINK NOT *-1 Nom BARRIER CC OR CLB OR CS OR CP);

Such as in "Mun vÃ¡lddÃ¡n duon girjji."

Changed the following rule to allow for topicalized objects in some cases:

>SELECT (Gen @GN>) IF (*-1 BOS LINK *1 (Gen @GN>) LINK 1 Nom LINK *1 TV)(0 N LINK 1 Nom LINK *1 TV LINK 0 VFIN);

<SELECT (Gen @GN>) IF (*-1 BOS LINK *1 (Gen @GN>) LINK 1 (N Nom) LINK *1 TV)(0 N LINK 1 Nom LINK *1 TV LINK 0 VFIN);

Added "bÃ¡lvalus" to LIST EVENT and "dieÄihit" to LIST ABSTRACT-ACTIVITY-VERB

Added EVENT to the following rules, assuming that EVENTs can have the same functions as HUMANs and would have a relational function if used in comitative case
and modified them
 with respect to a possible Aktio taking a preceeding locative adjunct:

>SELECT (Pl Loc) IF (*-1 (@SUBJ) LINK 0 HUMAN OR INSTITUTION OR (Pron Rel))(0 INSTITUTION OR HUMAN OR EVENT LINK 0 (Sg Com) LINK NOT -1 PRON-DU)(*1 ABSTRACT-ACTIVITY-VERB LINK NOT 0 Actio LINK NOT 1 N BARRIER CP OR CS OR COMMA);

>SELECT (Pl Loc) IF (*-1 ABSTRACT-ACTIVITY-VERB LINK NOT 0 Actio LINK NOT 1 N)(0 INSTITUTION OR HUMAN OR EVENT LINK 0 (Sg Com) BARRIER CP OR CS OR COMMA LINK NOT -1 PRON-DU);

Removed the REMOVE ACTIO rule
> #REMOVE Actio IF (0 PrfPrc)(1 N);
and restored a version close to the old one. We want to have something to the left of 0,
but this something shall neither be PRE-A-N, as in the first version, or nothing, as in
the version we now comment out. Rather, we make it "Dem or (Pron Indef)".
The logic is that an Actio construction is of the form "Gen Actio N", and that any determiner
in front of the putaqtive Actio points at it being a PrfPrc. The other possibility, of excluding PrfPrc if -1 Gen, is unfortunately not possible.
REMOVE Actio IF (-1 Dem OR (Pron Indef))(0 PrfPrc)(1 N);

Changed
< REMOVE PrfPrc IF (-1 NP-HEAD-GEN)(0 Actio)(1 N)(2 NP-BORDER OR Dem);
to
> SELECT Actio IF (-1 NP-HEAD-GEN)(0 Actio)(1 N)(2 NP-BORDER OR Dem);
i.e., changed REMOVE to SELECT. The resulting rule is strong, and should be monitored against a larger corpus.

Added the requirement "NOT Actio" to a number of rules.
> SELECT Gen IF (NOT *-1 ("leat") BARRIER CS)(NOT 0 Interr OR Dem)(1C Actio)
> SELECT (...)
> 	Dem OR Attr OR PrsPrc OR PrfPrc OR VFIN OR Actio)(NOT 0 TITLE LINK 1 Prop)
> SELECT Loc IF (NOT 0 ("dat") OR Actio)(NOT *1C Com BARRIER NPNH)

Strictly speaking, this might not be needed, given the stronger SELECT Actio rule above, but it hopefully will not do any harm either.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@7778 c7155fb1-f0a7-4240-a2fc-2600b6f42f90