Moved the following rule to the end of the @CS-COMPL chapter:
> ADD (@CS-COMPL) TARGET Nom IF (*-1 ("go") BARRIER NOT-NP LINK NOT -1 VFIN)
> 	(1 EOS);

Changed the following possessive rule by taking away the BARRIER:

< SELECT (PxSg3) IF (*-3 NP-HEAD-SG-NOM BARRIER V-NOT-AUX)(-2 A)(-1 go)

> SELECT (PxSg3) IF (*-3 (N Sg Nom) OR (Pron Pers Sg3 Nom)
>     )
> #    BARRIER V-NOT-AUX LINK NOT -1 AUX)
>     (-2C A)(-1 go)



Added ConNeg rule for (0 A) cases.
> REMOVE (Ind ConNeg) IF (-1 ConNeg)(*-2 (Neg Ind)
>     BARRIER VFIN OR ConNeg OR PrfPrc OR CS OR CP)(0 A)(1 go);
>     ## Máhttájeaddji ii leat stuorit go oahpaheaddjis, ii ge bálvaleaddji
>     ## stuorit go hearrás.
>     ## Máhttájeaddji ii leat stuorit go oahpaheaddjis, ii ge bálvaleaddji
>     ## stuorit johtileabbot go hearrás.


In the syntactic part, moved the whole chapter "# Verbal functions" from the end and almost to the beginning of the chapter. It seems we are better at selecting verbal functions, so let us do that in the beginning.