Added a constraint (NOT 1C Po) to the following rule in order to retain genitives preceeding sure postpositions:
>REMOVE Gen IF (*-1C Ord BARRIER NPNH OR Gen OR CC OR PUNCT OR REALCLB OR Ord
	LINK NOT 0 ("okta"))(NOT 0 VERB OR ("nr") OR ("nummar") OR ("s"))(NOT 1C Po)(NOT *1 N BARRIER NPNH);


Modified a couple of preposition-selection rules so that prepositions are not chosen if adverb interpretation is possible:
SELECT Pr IF (NOT -1 Gen OR (Pron Interr Nom))(NOT 0 A OR VFIN OR Adv)
	(*1 (N Gen) BARRIER NPNH LINK 0 HNOUN);


Added a	C to "*-1C Pr"

# Genitive after preposition
# - - - - - - - - - - - - - -
SELECT Gen IF (*-1C Pr BARRIER NPNH OR REALCLB OR CONTRA)(NOT 0 VFIN)
	(1 NP-BORDER OR CC OR COMMA OR ("ieÅ¡"));

SELECT Gen IF (*-1C Pr BARRIER NOT-NP OR Po)(*1 Interr BARRIER WORD);

SELECT Gen IF (-1C Pr)(0 Num);

and moved a copy of the the whole section to a later place in order to get another round of selection.

Solved the problem of locative or possessive suffix for the following sentence:

VuoÄÄun lea biddjon ahte dÃ¡t lanjadagat adnojit "boarrÃ¡siid-dikÅ¡osadjin", muhto maÅÅil teavsttas Äilget njealjÃ¡t Ã¡sodaga birra.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@7826 c7155fb1-f0a7-4240-a2fc-2600b6f42f90