Added rule:
!"Diphthong Simplification in Actor Nouns from u-Verbs"
!  Vx:0 <=> Vow _ Cns:+ u: Â¬Â§: [ X3: | X5: ] ;   ! Western and norm
!			where Vx in (e o a) ;

The rule is for the Divvun-project and must be commented in for the spell-checker.

Removed X2 and X6 and added X5 to rule:

"Diphthong Simplification in Actor Nouns from u-Verbs (Optional)"
  Vx:0 => Vow _ Cns:+ u: Â¬Â§: [ X3: | X5: ] ;   ! Eastern, substandard
			where Vx in (e o a) ;

The rule is for the Disamb. project and must be commented out for the spell-checker. Same idea for following rules:

!"Diphthong Simplification in Actor Nouns from i- and a-Stems
! vieh'ki > vihkkiin
!  Vx:0 <=> Vow _ Cns:+ [ i: | a: ] Â¬Â§: X5: ;			! Western and norm
!			where Vx in (e o a) ;

"Diphthong Simplification in Actor Nouns from i- and a-Stems (Optional)" ! vieh'ki > vihkkiin
  Vx:0 => Vow _ Cns:+ [ i: | a: ] Â¬Â§: X5: ;				  !	      viehkkiin
			where Vx in (e o a) ;				  ! Eastern, substandard
!âCÌ§Â¨ vieh'kiÂ¬Â§X5jn
!âCÌ§Â¨ vi0hkki00in
! Restored.


git-svn-id: https://gtsvn.uit.no/langtech/trunk/gt/sme@8047 c7155fb1-f0a7-4240-a2fc-2600b6f42f90